add description to file upload

+++ src/scripts/fileEmbedLink/upload.ts
@@ -1,4 +1,4 @@
-import { ref, onMounted, defineAsyncComponent } from 'vue';
+import { ref, onMounted, reactive } from 'vue';
 import { useRoute, useRouter } from "vue-router";
 import { mapState } from 'pinia';
 import { ElMessage } from 'element-plus';
@@ -6,6 +6,7 @@ import { useDataCategoryStore } from '@/stores/dataCategory';
 import { fileEmbedLinkApi } from '@/api/fileEmbedLinkApi';
 import { fileUploadApi } from '@/api/fileUploadApi';
 import { FileSelector, Dropzone, DialogButton } from 'vue3-file-selector';
+import type { FormInstance, FormRules } from 'element-plus';
 import { humanFileSize } from "@/helpers/ultilityFunctions";
 const appState = useDataCategoryStore();
 export default {
@@ -24,6 +25,23 @@ export default {
 
         const isLoading = ref(false);
         const currentOrganizationName = ref('');
+
+        const ruleFormValidationRef = ref<FormInstance>();
+        const rules = reactive<FormRules>({
+            description: [
+                {
+                    required: true,
+                    message: 'Vui lòng không bỏ trống..',
+                    trigger: 'blur',
+                },
+                {
+                    min: 3,
+                    message: 'Nhập tối thiểu 3 ký tự..',
+                    trigger: 'blur',
+                },
+            ],
+        });
+
         const currentEmbedLinkInfo = ref<any>(null);
         const itemModel = ref({
             organizationId: appState.defaultOrganization.id,
@@ -86,40 +104,47 @@ export default {
                 });
         };
 
-        const addItemSubmit = () => {
-            isLoading.value = true;
-            isUploadProgress.value = true;
-            try {
-                const formData = new FormData();
-                    formData.append('file', files.value[0]);
-                    let requestParams:any = { 
-                        organization_id : itemModel.value.organizationId,
-                        embedded_id: itemModel.value.embedLinkId,
-                    };
-                    fileUploadApi.uploadFileData(requestParams, formData, controllerUpload, onProgressUpload).then((response: any) => {
+        const addItemSubmit = async () => {
+            if (!ruleFormValidationRef || !ruleFormValidationRef.value) return;
+            await ruleFormValidationRef.value.validate((valid, fields) => {
+                if (valid) {                    
+                    isLoading.value = true;
+                    isUploadProgress.value = true;
+                    try {
+                        const formData = new FormData();
+                            formData.append('file', files.value[0]);
+                            let requestParams:any = { 
+                                organization_id : itemModel.value.organizationId,
+                                embedded_id: itemModel.value.embedLinkId,
+                            };
+                            fileUploadApi.uploadFileData(requestParams, formData, controllerUpload, onProgressUpload).then((response: any) => {
+                                isLoading.value = false;
+                                isUploadProgress.value = false;
+                                if(response){
+                                    switch(response.status){
+                                        case 200:						
+                                            ElMessage({
+                                                message: 'Đã tải lên file thành công !',
+                                                type: 'success',
+                                            });
+                                            removeFile();
+                                            break;
+                                    }
+                                }
+                            })
+                            .catch((response:any) =>{
+                                console.error('uploadFileData ERROR_1', response);
+                                ElMessage.error('Tải file lên không thành công. Đã có lỗi xảy ra !');
+                            });
+                    } catch (err) {
+                        console.log(err);
                         isLoading.value = false;
                         isUploadProgress.value = false;
-                        if(response){
-                            switch(response.status){
-                                case 200:						
-                                    ElMessage({
-                                        message: 'Đã tải lên file thành công !',
-                                        type: 'success',
-                                    });
-                                    removeFile();
-                                    break;
-                            }
-                        }
-                    })
-                    .catch((response:any) =>{
-                        console.error('uploadFileData ERROR_1', response);
-                        ElMessage.error('Tải file lên không thành công. Đã có lỗi xảy ra !');
-                    });
-            } catch (err) {
-                console.log(err);
-                isLoading.value = false;
-                isUploadProgress.value = false;
-            }
+                    }
+                } else {
+                    console.log('error submit!', fields);
+                }
+            });
         };
 
         const route:any = useRoute();
@@ -129,6 +154,8 @@ export default {
         });
         return {
             isLoading,
+            ruleFormValidationRef,
+            rules,
             itemModel,
             addItemSubmit,
             onOrganizationChanged,
